URGENT client.c ligne 57 : accès aux éléments du client (l.56) ; pb de pointeur

+++ structure.h
@@ -1,11 +1,11 @@
 
 typedef struct {
-	int type;
+	long type;
 	int etageDemande;
 	int etageAppuiBtn;
 } MessageEtageDemande;
 
 typedef struct {
-	int type;
+	long type;
 	int etage;
 } MessageEtage;
